replace mvc with web sample
