+++ netcore/sample/AspNetCore/Program.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace AspNetMvc
+namespace AspNetCore
 {
     public class Program
     {
