+++ netcore/sample/AspNetCore/Properties/launchSettings.json
@@ -3,8 +3,8 @@
     "windowsAuthentication": false, 
     "anonymousAuthentication": true, 
     "iisExpress": {
-      "applicationUrl": "http://localhost:19762",
-      "sslPort": 44367
+      "applicationUrl": "http://localhost:35206",
+      "sslPort": 44375
     }
   },
   "profiles": {
@@ -15,7 +15,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "AspNetMvc": {
+    "AspNetCore": {
       "commandName": "Project",
       "launchBrowser": true,
       "applicationUrl": "https://localhost:5001;http://localhost:5000",
