--- netcore/sample/AspNetMvc/Views/Home/Index.cshtml
@@ -1,8 +1,0 @@
-﻿@model HomeViewModel
-@{
-    ViewData["Title"] = "Home Page";
-}
-
-<div class="text-center">
-    <h3>CorLib: @Model.Info</h3>
-</div>
