--- netcore/sample/AspNetMvc/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using AspNetMvc
-@using AspNetMvc.Models
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
