--- netcore/sample/AspNetMvc/wwwroot/js/site.js
@@ -1,4 +1,0 @@
-﻿// Please see documentation at https://docs.microsoft.com/aspnet/core/client-side/bundling-and-minification
-// for details on configuring this project to bundle and minify static web assets.
-
-// Write your JavaScript code.
